Implement deploy workflow to camunda only
